Create user account management system for DBA

+++ src/main/java/com/el/spring/entity/User.java
@@ -9,7 +9,7 @@ public class User{
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
+    private int id;
 
     @Column(name = "username")
     private String username;
@@ -28,11 +28,11 @@ public class User{
     public User() {
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
